requirements.txt
loguru~=0.6.0
pycryptodome